装填に時間がかかるようにした他
カメラのコライダーを作成(仮)
経過時間のリセットを関数から砲弾クラスのUpdate内にした。
連射弾に生存時間を追加。

装填を開始してから弾を変更すると現在装備中の弾の方が装填されてしまうので、装填開始時に装填する弾を記憶する変数を作成する必要がある。

+++ Game/UserInterface/HpGauge.h
@@ -22,7 +22,7 @@ public:
 	const int GetDefaultValue() const { return DEFAULT_VALUE; }
 
 	// ƒ_ƒ[ƒW
-	void Damage(float damage) { m_value -= damage; }
+	void Damage(float damage) { m_value -= static_cast<int>(damage); }
 
 	// €–S‚µ‚Ä‚¢‚é‚©‚Ç‚¤‚©
 	bool GetDead() { return m_isDead; }
